[Silver III] Title: 주유소, Time: 368 ms, Memory: 36104 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -14,19 +14,19 @@ public class Main {
     public static void main(String[] args) throws IOException {
         int n = Integer.parseInt(br.readLine());
 
-        int[] distance = new int[n - 1];
-        int[] cost = new int[n];
-        int minCost = Integer.MAX_VALUE;
-        int result = 0;
+        long[] distance = new long[n - 1];
+        long[] cost = new long[n];
+        long minCost = Long.MAX_VALUE;
+        long result = 0;
 
         st = getSt();
         for (int i = 0; i < distance.length; i++) {
-            distance[i] = Integer.parseInt(st.nextToken());
+            distance[i] = Long.parseLong(st.nextToken());
         }
 
         st = getSt();
         for (int i = 0; i < n; i++) {
-            cost[i] = Integer.parseInt(st.nextToken());
+            cost[i] = Long.parseLong(st.nextToken());
         }
 
         for (int i = 0; i < n - 1; i++) {
